working on apple music search based on Stephen's work
