+++ Pipfile
@@ -16,6 +16,7 @@ django-cors-headers = "*"
 django-on-heroku = "*"
 gunicorn = "*"
 requests = "*"
+pyjwt = {extras = ["crypto"], version = "*"}
 
 [dev-packages]
 flake8 = "*"
