+++ api/apple_music_search.py
@@ -1,0 +1,37 @@
+import datetime
+import jwt
+
+
+secret = """-----BEGIN PRIVATE KEY-----
+MIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQgzWJ1tPxyB39nEjUn
+WCjymhCf2sfZF6IO3VJPNqpmRFOgCgYIKoZIzj0DAQehRANCAARE/4mbn57GPe8n
+/yd3zQ005COG7hj0A8L4f/4KoU4eDj1piuShHjdY3rle6TnAqw+srQBxjcCkvA2v
+692QycXe
+-----END PRIVATE KEY-----"""
+
+key_ID = "5BBHF2WLJF"
+team_ID = "C2NAAJY6Z3"
+alg = "ES256"
+
+time_now = datetime.datetime.now()
+time_expired = time_now + datetime.timedelta(hours=12)
+
+headers = {
+    "alg": alg,
+    "kid": key_ID
+}
+
+payload = {
+    "iss": team_ID,
+    "exp": int(time_expired.strftime("%s")),
+    "iat": int(time_now.strftime("%s"))
+}
+
+if __name__ == "__main__":
+    token = jwt.encode(payload, secret, algorithm=alg, headers=headers)
+
+    print("----TOKEN----")
+    print(token)
+
+    print("----CURL----")
+    print("curl -v -H 'Authorization: Bearer %s' \"https://api.music.apple.com/v1/catalog/us/artists/36954\" " % (token))
